moved collection,database,document into the database folder
