+++ include/NeoFOAM/finiteVolume/cellCentred/fieldCollection.hpp
@@ -10,9 +10,9 @@
 
 #include "NeoFOAM/core/demangle.hpp"
 #include "NeoFOAM/core/error.hpp"
-#include "NeoFOAM/core/database.hpp"
-#include "NeoFOAM/core/collection.hpp"
-#include "NeoFOAM/core/document.hpp"
+#include "NeoFOAM/core/database/database.hpp"
+#include "NeoFOAM/core/database/collection.hpp"
+#include "NeoFOAM/core/database/document.hpp"
 
 #include "NeoFOAM/finiteVolume/cellCentred/fields/volumeField.hpp"
