+++ include/NeoFOAM/finiteVolume/cellCentred/oldTimeCollection.hpp
@@ -9,9 +9,9 @@
 #include "NeoFOAM/core/demangle.hpp"
 #include "NeoFOAM/core/error.hpp"
 
-#include "NeoFOAM/core/database.hpp"
-#include "NeoFOAM/core/collection.hpp"
-#include "NeoFOAM/core/document.hpp"
+#include "NeoFOAM/core/database/database.hpp"
+#include "NeoFOAM/core/database/collection.hpp"
+#include "NeoFOAM/core/database/document.hpp"
 
 #include "NeoFOAM/finiteVolume/cellCentred/fields/volumeField.hpp"
 #include "NeoFOAM/finiteVolume/cellCentred/fieldCollection.hpp"
