+++ src/CMakeLists.txt
@@ -11,9 +11,9 @@ target_sources(
   PRIVATE "core/mpi/halfDuplexCommBuffer.cpp"
           "core/primitives/vector.cpp"
           "core/time.cpp"
-          "core/collection.cpp"
-          "core/database.cpp"
-          "core/document.cpp"
+          "core/database/database.cpp"
+          "core/database/collection.cpp"
+          "core/database/document.cpp"
           "core/dictionary.cpp"
           "core/demangle.cpp"
           "core/tokenList.cpp"
