+++ src/core/database/collection.cpp
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: MIT
 // SPDX-FileCopyrightText: 2024 NeoFOAM authors
 
-#include "NeoFOAM/core/collection.hpp"
+#include "NeoFOAM/core/database/collection.hpp"
 
 
 namespace NeoFOAM
