+++ test/core/CMakeLists.txt
@@ -3,10 +3,8 @@
 
 add_subdirectory(mpi)
 add_subdirectory(primitives)
+add_subdirectory(database)
 
-neofoam_unit_test(collection)
-neofoam_unit_test(database)
-neofoam_unit_test(document)
 neofoam_unit_test(dictionary)
 neofoam_unit_test(tokenList)
 neofoam_unit_test(inputs)
