+++ test/core/database/CMakeLists.txt
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Unlicense
+# SPDX-FileCopyrightText: 2024 NeoFOAM authors
+
+neofoam_unit_test(collection)
+neofoam_unit_test(database)
+neofoam_unit_test(document)
