+++ test/core/database/collection.cpp
@@ -7,9 +7,9 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators_all.hpp>
 
-#include "NeoFOAM/core/database.hpp"
-#include "NeoFOAM/core/collection.hpp"
-#include "NeoFOAM/core/document.hpp"
+#include "NeoFOAM/core/database/database.hpp"
+#include "NeoFOAM/core/database/collection.hpp"
+#include "NeoFOAM/core/database/document.hpp"
 
 #include "customTestCollection.hpp"
