+++ test/core/database/document.cpp
@@ -7,7 +7,7 @@
 #include <catch2/catch_test_macros.hpp>
 #include <catch2/generators/catch_generators_all.hpp>
 
-#include "NeoFOAM/core/document.hpp"
+#include "NeoFOAM/core/database/document.hpp"
 
 
 TEST_CASE("Document")
