+++ test/finiteVolume/cellCentred/fields/fieldCollection.cpp
@@ -10,7 +10,7 @@
 #include "NeoFOAM/finiteVolume/cellCentred/fieldCollection.hpp"
 #include "NeoFOAM/finiteVolume/cellCentred/fields/volumeField.hpp"
 #include "NeoFOAM/finiteVolume/cellCentred/boundary/volumeBoundaryFactory.hpp"
-#include "NeoFOAM/core/database.hpp"
+#include "NeoFOAM/core/database/database.hpp"
 
 namespace fvcc = NeoFOAM::finiteVolume::cellCentred;
